fix: исправить роутинг, убрать дубликат pages
