+++ src/pages/project-details/ProjectDetails.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Link, useParams } from "react-router-dom";
-import { routesPaths } from "../providers/router/routesPaths";
+import { routesPaths } from "../../app/providers/router/routesPaths";
 
 const ProjectDetails: React.FC = () => {
   const { slug } = useParams<{ slug: string }>();
